Move Classpaths into its own file
